+++ libc/src/sched/linux/sched_setparam.cpp
@@ -18,7 +18,8 @@ namespace __llvm_libc {
 
 LLVM_LIBC_FUNCTION(int, sched_setparam,
                    (pid_t tid, const struct sched_param *param)) {
-  long ret = __llvm_libc::syscall_impl(SYS_sched_setparam, tid, param);
+  int ret = static_cast<int>(
+      __llvm_libc::syscall_impl(SYS_sched_setparam, tid, param));
   if (ret < 0) {
     libc_errno = -ret;
     return -1;
