+++ libc/src/unistd/linux/lseek.cpp
@@ -21,7 +21,8 @@ namespace __llvm_libc {
 LLVM_LIBC_FUNCTION(off_t, lseek, (int fd, off_t offset, int whence)) {
   off_t result;
 #ifdef SYS_lseek
-  long ret = __llvm_libc::syscall_impl(SYS_lseek, fd, offset, whence);
+  int ret = static_cast<int>(
+      __llvm_libc::syscall_impl(SYS_lseek, fd, offset, whence));
   result = ret;
 #elif defined(SYS_llseek)
   long ret = __llvm_libc::syscall_impl(SYS_llseek, fd,
@@ -29,8 +30,8 @@ LLVM_LIBC_FUNCTION(off_t, lseek, (int fd, off_t offset, int whence)) {
                                        (long)offset, &result, whence);
   result = ret;
 #elif defined(SYS__llseek)
-  long ret = __llvm_libc::syscall_impl(SYS__llseek, fd, offset >> 32, offset,
-                                       &result, whence);
+  int ret = static_cast<int>(__llvm_libc::syscall_impl(
+      SYS__llseek, fd, offset >> 32, offset, &result, whence));
 #else
 #error "lseek, llseek and _llseek syscalls not available."
 #endif
